requirements.txt
python-telegram-bot==21.4
SQLAlchemy==2.0.29
aiosqlite==0.20.0
asyncpg==0.29.0
Pillow==10.3.0
httpx==0.27.0
pandas==2.2.1
matplotlib==3.8.4
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn[standard]==0.30.3
